#37 :: Move getRandomBoard code into globals, fix logic error, and add test case
