+++ src/state/boardSlice.ts
@@ -2,9 +2,9 @@ import { CaseReducer, createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { NUM_COLS, NUM_ROWS } from '../constants';
 import {
   addBlock as addBlockHelper,
-  Block,
   getMinPos,
   getNewPosFromDir,
+  getRandomBoard,
   Grid,
   Move,
   moveBlock as moveBlockHelper,
@@ -25,37 +25,6 @@ const initialState: BoardState = {
   grid: Array(NUM_ROWS)
     .fill(null)
     .map(() => Array(NUM_COLS).fill(0)),
-};
-
-//  Randomize Helpers
-
-const getRandomAvailableCoords = (grid: Grid): Pos => {
-  const getRandomCoords = () => [
-    Math.floor(Math.random()*NUM_ROWS),
-    Math.floor(Math.random()*NUM_COLS),
-  ];
-  let [i, j] = getRandomCoords();
-  while (grid[i][j] !== 0) {
-    [i, j] = getRandomCoords();
-  }
-  return { row: i, col: j };
-};
-
-const getRandomBlock = (
-  numCellsAvailable: number, hasTwoByTwoBlock: boolean
-): Block => {
-  let availableBlocks: Block[] = [
-    { rows: 1, cols: 1 },
-    { rows: 2, cols: 1 },
-    { rows: 1, cols: 2 },
-    { rows: 2, cols: 2 },
-  ];
-  if (numCellsAvailable === 1) {
-    availableBlocks = [{ rows: 1, cols: 1 }];
-  } else if (!hasTwoByTwoBlock) {
-    availableBlocks = [{ rows: 2, cols: 2 }];
-  }
-  return availableBlocks[Math.floor(Math.random() * availableBlocks.length)];
 };
 
 // Actions
@@ -84,27 +53,7 @@ const moveBlockToPosReducer: CaseReducer<
 );
 
 const randomizeReducer: CaseReducer<BoardState> = (state) => {
-  let numCellsAvailable = NUM_ROWS*NUM_COLS-2;
-  let hasTwoByTwoBlock = false;
-
-  while (numCellsAvailable > 0) {
-    const randomPosBlock = {
-      block: getRandomBlock(numCellsAvailable, hasTwoByTwoBlock),
-      pos: getRandomAvailableCoords(state.grid),
-    };
-    // attempt to add random block at random pos
-    try {
-      addBlockHelper(state, randomPosBlock);
-    } catch {
-      continue;
-    }
-    // if successful update variables
-    const blockArea = randomPosBlock.block.rows * randomPosBlock.block.cols;
-    numCellsAvailable -= blockArea;
-    if (blockArea === 4) {
-      hasTwoByTwoBlock = true;
-    }
-  }
+  getRandomBoard(state);
 };
 
 const removeBlockReducer: CaseReducer<
